requirements.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

droid-django-dajaxice
droid-common-js

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.
